% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pboost4sar.R
\name{EBIC.Sarlm}
\alias{EBIC.Sarlm}
\title{Extended BIC for SAM}
\usage{
\method{EBIC}{Sarlm}(object, p, p.keep, ...)
}
\arguments{
\item{object}{See \link[pboost:EBIC]{pboost::EBIC}.}

\item{p}{See \link[pboost:EBIC]{pboost::EBIC}.}

\item{p.keep}{See \link[pboost:EBIC]{pboost::EBIC}.}

\item{...}{See \link[pboost:EBIC]{pboost::EBIC}.}
}
\value{
EBIC value.
}
\description{
EBIC for spatial auto-regressive model.
}
