% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward-sar.R
\name{forward_sar}
\alias{forward_sar}
\title{Forward Stepwise Spatial Auto-Regressive Model}
\usage{
forward_sar(x, y, w)
}
\arguments{
\item{x}{Matrix of covariates.}

\item{y}{Vector of response.}

\item{w}{Weight matrix (row-sum scaled being one).}
}
\value{
Model object fitted on the selected features.
}
\description{
Forward stepwise strategy for spatial auto-regressive model.
}
\examples{
library(spdep)
library(spatialreg)

set.seed(2025)
b0 <- c(1.5, 3.0, 2.0, rep(0.0, 3))
rho0 <- 0.2
sig0 <- 1.0
n <- 81

DF <- simu_sar_data_rook(b0, rho0, sig0, n, sqrt(n), sqrt(n))
x <- DF[["x"]]
y <- DF[["y"]]
w <- DF[["w"]]

system.time( egg <- forward_sar(x, y, w) )
y.tilde <- (diag(NROW(x)) - egg[["rho"]] * w) \%*\% y

flag <- egg[["flag"]]
beta.hat <- egg[["beta"]][sort(names(egg[["beta"]]))]
sig2.hat <- mean( (y.tilde - drop(x[, flag, drop=FALSE] \%*\% beta.hat))^2 )
print( egg[["sig2"]] - sig2.hat )

}
