% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rho.R
\name{get_rho}
\alias{get_rho}
\title{MLE of \eqn{\rho}}
\usage{
get_rho(x, y, w)
}
\arguments{
\item{x}{Covariate matrix.}

\item{y}{Response vector.}

\item{w}{Weight matrix (row-sum scaled).}
}
\value{
MLE of \eqn{\rho}.
}
\description{
MLE of \eqn{\rho}.
}
\examples{
set.seed(2025)
b0 <- c(1.5, 3.0, 2.0, rep(0.0, 3))
rho0 <- 0.2
sig0 <- 1.0
n <- 81

DF <- simu_sam_data_rook(b0, rho0, sig0, n)
system.time( get_rho(as.matrix(DF[["X"]]), DF[["y"]], DF[["W0"]]) )

}
