% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pboost4sar-package.R
\docType{package}
\name{pboost4sar-package}
\alias{pboost4sar}
\alias{pboost4sar-package}
\title{pboost4sar: Profiled Variable Selection (profile boosting) for Spatial Auto-Regressive Model}
\description{
Profiled variable selection (profile boosting) for spatial auto-regressive (SAR) model.
}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com}

}
\keyword{internal}
