% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pboostsam-package.R
\docType{package}
\name{pboostsam-package}
\alias{pboostsam}
\alias{pboostsam-package}
\title{pboostsam: Profile Boosting Variable Selection for Spatial Autoregressive Model}
\description{
Profile boosting variable selection for spatial autoregressive model.
}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com}

}
\keyword{internal}
