% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pboostsam.R
\name{pboostsam}
\alias{pboostsam}
\title{Profile Boosting Feature Selection for Spatial Autoregressive Model}
\usage{
pboostsam(
  formula,
  data = list(),
  listw,
  na.action,
  Durbin,
  type,
  method = "eigen",
  quiet = NULL,
  zero.policy = NULL,
  interval = NULL,
  tol.solve = .Machine$double.eps,
  trs = NULL,
  control = list(),
  stopFun = EBIC,
  keep = NULL,
  maxK = NULL,
  verbose = FALSE
)
}
\arguments{
\item{formula}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{data}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{listw}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{na.action}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{Durbin}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{type}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{method}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{quiet}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{zero.policy}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{interval}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{tol.solve}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{trs}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{control}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{stopFun}{Stopping rule for profile boosting, which has the form
stopFun(object) to evaluate the performance of model object returned
by fitFun, such as \link[pboost:EBIC]{pboost::EBIC} or \link{BIC}.}

\item{keep}{Initial set of features that are included in model fitting.
\strong{If \code{keep} is specified, it should also be fully included in the RHS of \code{formula}.}}

\item{maxK}{Maximal number of identified features.
If \code{maxK} is specified, it will supress \code{stopFun}, saying that the
profile boosting continues until the procedure identifies \code{maxK} features.
The pre-specified features in \code{keep} are counted toward \code{maxK}.}

\item{verbose}{Print the procedure path?}
}
\value{
Model object fitted on the selected features.
}
\description{
Profile Boosting Feature Selection for Spatial Autoregressive Model
}
\examples{
library(spdep)
library(spatialreg)

set.seed(2025)
b0 <- c(1.5, 3.0, 2.0, rep(0.0, 3))
rho0 <- 0.2
sig0 <- 1.0
n <- 81

DF <- simu_sam_data_rook(b0, rho0, sig0, n)

data <- with(DF, data.frame(y, X))
W0 <- DF[["W0"]]
( result <- pboostsam(y ~ ., data, mat2listw(W0, style="W"), verbose=TRUE) )


}
