% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs4sar-package.R
\docType{package}
\name{pvs4sar-package}
\alias{pvs4sar-package}
\title{pvs4sar: The Profiled Variable Selection for Spatial Auto-Regressive Model}
\description{
Profiled variable selection for spatial auto-regressive model.
}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com}

}
\keyword{internal}
