% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs4sar.R
\name{pvs4sar}
\alias{pvs4sar}
\alias{pvs4sar_lagsarlm}
\title{Profiled Variable Selection for Spatial Auto-regressive Model}
\usage{
pvs4sar(x, y, w, maxK = NULL, keep = NULL, verbose = FALSE)

pvs4sar_lagsarlm(
  formula,
  data = list(),
  listw,
  na.action,
  Durbin,
  type,
  method = "eigen",
  quiet = NULL,
  zero.policy = NULL,
  interval = NULL,
  tol.solve = .Machine$double.eps,
  trs = NULL,
  control = list(),
  stopFun = EBIC,
  keep = NULL,
  maxK = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{Matrix of covariates.}

\item{y}{Vector of response.}

\item{w}{Weight matrix (row-sum scaled being one).}

\item{maxK}{Maximal number of identified features.
If \code{maxK} is specified, it will supress \code{stopFun}, saying that the
profile boosting continues until the procedure identifies \code{maxK} features.
The pre-specified features in \code{keep} are counted toward \code{maxK}.
Also see \link[pboost:pboost]{pboost::pboost}.}

\item{keep}{Initial set of features that are included in model fitting.
\strong{If \code{keep} is specified, it should also be fully included in the RHS of \code{formula}.}
Also see \link[pboost:pboost]{pboost::pboost}.}

\item{verbose}{Print the procedure path? Also see \link[pboost:pboost]{pboost::pboost}.}

\item{formula}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{data}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{listw}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{na.action}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{Durbin}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{type}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{method}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{quiet}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{zero.policy}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{interval}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{tol.solve}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{trs}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{control}{Parameters passed to \link[spatialreg:ML_models]{spatialreg::lagsarlm}.}

\item{stopFun}{Stopping rule for profile boosting, which has the form
stopFun(object) to evaluate the performance of model object returned
by fitFun, such as \link[pboost:EBIC]{pboost::EBIC} or \link{BIC}.}
}
\value{
Model object fitted on the selected features.
}
\description{
Profile boosting variable selection for spatial auto-regressive (SAR) model
\deqn{\bm{y} = \rho W \bm{y} + X \beta + \bm{\varepsilon},}
where \eqn{\bm{y}, \bm{\varepsilon} \in \mathbb{R}^n}, \eqn{X \in \mathbb{R}^{n \times p}},
\eqn{\rho \in (-1, 1)}.

\code{pvs4sar_lagsarlm()} equips the PVS procedure for \link[spatialreg:ML_models]{spatialreg::lagsarlm}.
}
\examples{
library(spdep)
library(spatialreg)

set.seed(2025)
b0 <- c(1.5, 3.0, 2.0, rep(0.0, 3))
rho0 <- 0.2
sig0 <- 1.0
n <- 81

DF <- simu_sar_data_rook(b0, rho0, sig0, n, sqrt(n), sqrt(n))
x <- DF[["x"]]
y <- DF[["y"]]
w <- DF[["w"]]

system.time( egg <- pvs4sar(x, y, w, verbose=TRUE) )
y.tilde <- (diag(NROW(x)) - egg[["rho"]] * w) \%*\% y

flag <- egg[["flag"]]
beta.hat <- egg[["beta"]][sort(names(egg[["beta"]]))]
sig2.hat <- mean( (y.tilde - drop(x[, flag, drop=FALSE] \%*\% beta.hat))^2 )
print( egg[["sig2"]] - sig2.hat )

system.time(
 pvs4sar_lagsarlm(y ~ ., data.frame(y, x), mat2listw(DF[["w"]], style="W"), verbose=TRUE)
)

}
