% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam-adaptivelasso.R
\name{sam_adaptivelasso}
\alias{sam_adaptivelasso}
\alias{tune_sam_adaptivelasso}
\alias{sam-adaptivelasso}
\title{Adaptive Lasso for Spatial Autoregressive Model}
\usage{
sam_adaptivelasso(X, Y, W, rho, lambda, max.iter = 1000, tol = 1e-06)

tune_sam_adaptivelasso(
  X,
  Y,
  W,
  rho,
  lambda.vec = 10^seq(-1, 1, length = 100),
  max.iter = 1000,
  tol = 1e-06
)
}
\arguments{
\item{X}{Covariate matrix.}

\item{Y}{Response vector.}

\item{W}{Weight matrix.}

\item{rho}{Auto-correlation coefficient.}

\item{lambda}{Tuning parameter.}

\item{max.iter}{Maximal number of iterations.}

\item{tol}{Covergence tolerance.}

\item{lambda.vec}{Vector of \eqn{\lambda}.}
}
\value{
\itemize{
\item \code{sam_adaptivelasso}: \code{list(beta, sig2, rho, flag, BIC)}.
\item \code{tune_sam_adaptivelasso}: \code{list(beta, sig2, rho, flag, BIC, lambda.opt)}.
}
}
\description{
Adaptive lasso for spatial autoregressive model.
}
\examples{
set.seed(2025)
b0 <- c(1.5, 3.0, 2.0, rep(0.0, 3))
rho0 <- 0.2
sig0 <- 1.0
n <- 81

DF <- simu_sam_data_rook(b0, rho0, sig0, n)
Y <- DF[["y"]]
X <- as.matrix(DF[["X"]])
W0 <- DF[["W0"]]

system.time( rho.hat <- get_rho(X, Y, W0) )
system.time( tune_sam_adaptivelasso(X, Y, W0) )

}
