% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu-sam-data.R
\name{simu_sam_data_rook}
\alias{simu_sam_data_rook}
\title{Simulate Data}
\usage{
simu_sam_data_rook(b0, rho0, sig0, n)
}
\arguments{
\item{b0}{True vector of \eqn{\beta_0}.
\itemize{
\item \code{c(1.5, 3.0, 2.0, rep(0.0, 3))}
\item \code{c(rep(1.5, 3), rep(2.0, 4), rep(3.0, 3), rep(0.0, 10))}
\item \code{c(rep(1.5, 5), rep(2.0, 5), rep(3.0, 5), rep(0.0, 15))}
}}

\item{rho0}{\eqn{\rho_0}.}

\item{sig0}{\eqn{\sigma_0} for \eqn{\varepsilon}.}

\item{n}{Sample size.}
}
\description{
Rook weight matrix.
}
