% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve-ridge.R
\name{solve_ridge}
\alias{solve_ridge}
\title{Ridge-Type Estimation}
\usage{
solve_ridge(X, y, d)
}
\arguments{
\item{X}{Matrix of size \eqn{n \times p}.}

\item{y}{Vector of length \eqn{n}.}

\item{d}{Positive scalar or vector.}
}
\value{
Vector of length \eqn{p}.
}
\description{
Solve the linear system \eqn{(X^\top X + D) \beta = y}.
Using Woodbury matrix identity
\eqn{(D + X^\top X)^{-1} = D^{-1} - D^{-1} X^\top (I_n + X D^{-1} X'^\top)^{-1} X^\top D^{-1}},
we have
\deqn{(D + X^\top X)^{-1} X^\top Y = D^{-1} X^\top (I_n + X D^{-1} X^\top)^{-1} y.}
Specifically,
\deqn{(X^\top X + \lambda I_p)^{-1} X^\top y = X^\top (XX^\top + \lambda I_n)^{-1} y.}
}
\examples{
set.seed(2026)
n <- 100
p <- 500
X <- matrix(rnorm(n * p), n, p)
y <- rnorm(n)

system.time( b1 <- solve_ridge(X, y, 0.1) )
max(abs(b1 - solve(crossprod(X) + 0.1 * diag(p), crossprod(X, y))))

d <- runif(p, 0.01, 0.5)
system.time( b2 <- solve_ridge(X, y, d) )
max(abs(b2 - solve(crossprod(X) + diag(d), crossprod(X, y))))

}
